changed name of application

+++ android/src/components/App.js
@@ -15,7 +15,14 @@ import Realm from 'realm';
 import Schema from '../realm';
 import Firebase from '../../../includes/firebase/firebase';
 
-Firebase.initialize();
+import Firestack from 'react-native-firestack';
+
+console.log('initialize')
+//Firebase.initialize();
+// const firestack = new Firestack({
+//   debug: true
+// });
+console.log('after initialize')
 
 
 const AppNavigator = DrawerNavigator({
@@ -67,16 +74,31 @@ export default class App extends Component {
 
 
     let timerLists = this.realm.objects('Timers');
-    if (timerLists.length > 1) { // Initializing Timers automatically gives it a length of 1 with an empty list object
+    let ticketList = this.realm.objects('Ticketed');
+    let expiredList = this.realm.objects('Expired');
+    if (timerLists.length >= 1) { // Initializing Timers automatically gives it a length of 1 with an empty list object
       let i = 0, lastTime;
-      while (lastTime === undefined && timerLists[i + 1]) { // Edge case for empty first object
-        if (timerLists[i].list.length > 0) lastTime = timerLists[i].list[0].createdAt;
+      while (lastTime === undefined && (timerLists[i + 1] || ticketList[0].list[i + 1] || expiredList[0].list[i + 1])) { // Edge case for empty first object
+        if (timerLists[i].list.length > 0) {
+          lastTime = timerLists[i].list[0].createdAt;
+          break;
+        }
+        if (ticketList[0].list.length > 0) {
+          lastTime = ticketList[0].list[i].createdAt;
+          break;
+        }
+        if (expiredList[0].list.length > 0) {
+          lastTime = expiredList[0].list[i].createdAt;
+          break;
+        }
         i++;
       }
       let context = this;
       let now = new Date();
-      if (now - lastTime > 28800000) { // Reset DB after 8 hours of activity
+      if (!lastTime || now - lastTime > 28800000) { // Reset DB after 8 hours of activity
         this._loopDeletion(timerLists);
+        this._loopDeletion(ticketList, true);
+        this._loopDeletion(expiredList, true);
         console.log('deletion')
         //TODO Doesn't wait..
         setTimeout(() => {
@@ -105,9 +127,9 @@ export default class App extends Component {
     console.log('signed in user!');
   };
 
-  _loopDeletion(timerLists) {
-    timerLists.forEach((timerList, idx) => {
-      timerList.list.forEach((timer, sidx) => {
+  _loopDeletion(timerLists, once) {
+    if (once) {
+      timerLists[0].list.forEach((timer, idx) => {
         RNFS.unlink(timer.mediaPath)
         .then(() => {
           console.log('FILE DELETED');
@@ -115,12 +137,28 @@ export default class App extends Component {
           .then(() => {
             console.log('PICTURE REMOVED');
             this.realm.write(() => {
-              //timerLists[idx]['list'].pop();
-              this.realm.delete(timerLists[idx]['list'][sidx]);
+              this.realm.delete(timerLists[0]['list'][idx]);
             });
           });
         });
       });
-    });
+    } else {
+      timerLists.forEach((timerList, idx) => {
+        timerList.list.forEach((timer, sidx) => {
+          RNFS.unlink(timer.mediaPath)
+          .then(() => {
+            console.log('FILE DELETED');
+            RNFS.exists(timer.mediaUri)
+            .then(() => {
+              console.log('PICTURE REMOVED');
+              this.realm.write(() => {
+                //timerLists[idx]['list'].pop();
+                this.realm.delete(timerLists[idx]['list'][sidx]);
+              });
+            });
+          });
+        });
+      });
+    }
   }
 }
